Log backtraces if present when jobs fail.
This is already done on sneakers master, but by logging each line of the
backtrace on a separate line. While this looks nice, it makes it very
hard for log tools to get the entire backtrace by searching for lines
that contain "ERROR" for instance. I'll probably merge sneakers master
into our fork and undo that change so we are all sync'd up again.