Further tweaks for performance ...
I was not really happy that trying to short circuit
and save a few method calls when the feature is
not in use only to impact on the feature itself.

This makes the whole thing a little more efficient
by switching to attr_reader and imposes a constant
cost whichever execution path is followed.