Outline of support for fatal (non-retriable) errors
This commit is an outline of how support for fatal (non-retriable)
errors could be integrated into sneakers in a backwards-compatible way.
This is useful for handling messages that can never succeed (e.g.
poison messages) which at the moment will be retried in the Maxretry
handler rather than going straight to the error queue.

Note that although this is (I believe) a functionally complete
implementation the tests are not complete, as I want to gauge interest
and reaction to this before spending too much time on it.