Ability to pass 'properties' to Bunny session
With this change we will be able to define a `properties` property in
the `Sneakers` configuration, that will be passed to `Bunny` when a
connection is created.

```ruby
Sneakers.configure({
  amqp: 'amqp://guest:guest@localhost:5672',
  properties: {
    key: "value",
    key2: "value2"
  }
})
```

This can be useful, for example, to identify where a connection is
coming from, when we have dozens of applications using a `RabbitMQ` instance.

This is something that can already be done with `Bunny`:

```ruby
Bunny.new(connection_string, { properties: { key: 'value' } })
```

So we are just leveraging that functionality when the connection is
created from `Sneakers`.