BREAKING CHANGE: remove enforced Worker#work timeout
See #343 for the background. Worker implementations now must handle
timeouts in a way they see fit. This is unfortunate but there is no
way to enforce timeouts in a safe way without at least some Worker implementation
cooperation and currently there is no decision on what that should look like.

We previously increased the timeout to 10 minutes. This is about as good
as not having any, so let's rip off the bandaid.

Closes #343.